requirements.txt
# Sphinx
sphinx<8
sphinx-book-theme==1.1.4
myst-parser==4.0.1
sphinx-design==0.6.1
unidecode==1.3.8
sphinxcontrib-images==0.9.4
sphinxcontrib-plantuml==0.30
sphinx-copybutton==0.2.10
linkify-it-py==2.0.3
setuptools>=45.0.0

# FastAPI
fastapi == 0.115.11
uvicorn == 0.34.0
passlib == 1.7.4
python-jose == 3.4.0
sqlalchemy == 2.0.39
asyncpg == 0.30.0
databases == 0.9.0
alembic == 1.15.1
dotenv == 0.9.9

# Установка локальных расширений для Sphinx
-e ./localextensions/sphinxcontrib-images
-e ./example-plate
